requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
passlib==1.7.4
itsdangerous==2.2.0
paho-mqtt==1.6.1